Update rht.py
Changed line 495 datatype of array. This line creates an empty array and fills it with nans, but originally the datatype was int and python will not fill an int array with nans so reports an error. I have changed the datatype of the empty array to float so it can be filled with nans.